requirements.txt
aiohttp                ==3.8.1
aiosignal              ==1.2.0
albumentations         ==1.1.0
async-timeout          ==4.0.2
attrs                  ==21.2.0
autobahn               ==22.3.2
Automat                ==20.2.0
av                     ==8.1.0
certifi                ==2021.10.8
cffi                   ==1.15.0
charset-normalizer     ==2.0.9
click                  ==8.0.3
colorama               ==0.4.4
comet-ml               ==3.23.0
configobj              ==5.0.6
constantly             ==15.1.0
cryptography           ==36.0.2
cycler                 ==0.11.0
Deprecated             ==1.2.13
dulwich                ==0.20.26
einops                 ==0.4.0
et-xmlfile             ==1.1.0
everett                ==2.0.1
fonttools              ==4.37.1
frozenlist             ==1.3.0
fvcore                 ==0.1.5.post20220119
humanize               ==3.13.1
hyperlink              ==21.0.0
idna                   ==3.3
imageio                ==2.13.3
importlib-resources    ==5.4.0
incremental            ==21.3.0
iopath                 ==0.1.9
Jinja2                 ==3.1.1
joblib                 ==1.1.0
jsonschema             ==4.2.1
kiwisolver             ==1.3.2
MarkupSafe             ==2.1.1
matplotlib             ==3.5.1
mkl-fft                ==1.3.1
mkl-random             ==1.2.2
mkl-service            ==2.4.0
monai                  ==0.9.0
multidict              ==6.0.2
networkx               ==2.6.3
nibabel                ==3.2.1
nvidia-ml-py3          ==7.352.0
olefile                ==0.46
opencv-python          ==4.5.4.60
opencv-python-headless ==4.5.4.60
openpyxl               ==3.0.9
packaging              ==21.3
pandas                 ==1.3.5
parameterized          ==0.8.1
patsy                  ==0.5.2
Pillow                 ==8.4.0
pip                    ==21.2.2
portalocker            ==2.3.2
pycparser              ==2.21
pydicom                ==2.2.2
PyOpenGL               ==3.1.6
pyparsing              ==3.0.6
pyrsistent             ==0.18.0
python-dateutil        ==2.8.2
pytorchvideo           ==0.1.5
pytorch                ==1.12.1
pytz                   ==2021.3
PyWavelets             ==1.2.0
PyYAML                 ==6.0
pyzmq                  ==22.3.0
qudida                 ==0.0.4
requests               ==2.26.0
requests-toolbelt      ==0.9.1
scikit-image           ==0.19.0
scikit-learn           ==1.0.1
scipy                  ==1.7.3
seaborn                ==0.11.2
semantic-version       ==2.8.5
setuptools             ==58.0.4
SimpleITK              ==2.1.1
six                    ==1.16.0
statsmodels            ==0.13.2
tabulate               ==0.8.9
termcolor              ==1.1.0
threadpoolctl          ==3.0.0
tifffile               ==2021.11.2
torch-geometric        ==2.0.4
torchio                ==0.18.71
tqdm                   ==4.62.3
Twisted                ==22.2.0
txaio                  ==22.2.1
typing-extensions      ==3.10.0.2
urllib3                ==1.26.7
vedo                   ==2022.1.0
vtk                    ==9.0.3
websocket-client       ==1.2.3
wheel                  ==0.37.0
wincertstore           ==0.2
wrapt                  ==1.13.3
wslink                 ==1.4.3
wurlitzer              ==3.0.2
xgboost                ==1.7.3
yacs                   ==0.1.8
yarl                   ==1.7.2
zipp                   ==3.6.0
zope.interface         ==5.4.0
